Naprawiono problem z rekrutacją Bigginsa. Naprawiono problem z podkładaniem ładunków pod wentylator. (Sektor J13). Naprawiono przejście z sektora K14-b1 do K15-b1. Naprawiono problem z dialogami Jerrego. Możliwość włączenia\wyłączenia strony Bobby Ray Możliwość zmiany sektora kończącego grę. Możliwość włączenia\wyłączenia desantu z helikoptera lub włączenia\wyłączenia katastrofy śmigłowca. Możliwość włączenia\wyłączenia ubezpieczenia najemników na stronie www. Modyfikacja skryptu „GameInit.lua”. Dodano nową  funkcje „InitStrategicLayer”. Funkcja służy do inicjowania sektora początkowego. Uwaga! Zmian sektora z H7 na inny nie działa poprawnie! Można zmieniać położenie katastrofy tylko w sektorze H7.

+++ GameSettings.h
@@ -909,8 +909,8 @@ typedef struct
 	UINT16 usHelicopterBaseCostPerRedTile;
 
 	// HEADROCK HAM 3.5: Experimental, arrival sectorX/Y.
-	UINT8 ubDefaultArrivalSectorX;
-	UINT8 ubDefaultArrivalSectorY;
+	INT16 ubDefaultArrivalSectorX;
+	INT16 ubDefaultArrivalSectorY;
 
 	// HEADROCK HAM 3.5: Limit bonus from tracers based on range to target. This is a divisor for the effect.
 	UINT8 ubRangeDifficultyAimingWithTracers;
